+++ src/lib/superValidate.ts
@@ -137,6 +137,8 @@ export async function superValidate<
 	};
 }
 
+/////////////////////////////////////////////////////////////////////
+
 /**
  * Sends a message with a form, with an optional HTTP status code that will set
  * form.valid to false if status >= 400. A status lower than 400 cannot be sent.