+++ src/tests/superValidate.test.ts
@@ -1,6 +1,6 @@
 import type { JSONSchema } from '$lib/jsonSchema/index.js';
 import { describe, it, expect } from 'vitest';
-import { toJsonSchema, type ValidationAdapter } from '$lib/adapters/index.js';
+import { type ValidationAdapter } from '$lib/adapters/index.js';
 import { Foo, bigZodSchema } from './data.js';
 import { constraints } from '$lib/jsonSchema/constraints.js';
 import { defaultValues } from '$lib/jsonSchema/defaultValues.js';
@@ -13,7 +13,7 @@ import { zod, zodToJsonSchema } from '$lib/adapters/zod.js';
 import { z } from 'zod';
 
 import { valibot } from '$lib/adapters/valibot.js';
-import { object, string, email, minLength, array } from 'valibot';
+import { object, string, email, minLength, array, integer, number, minValue } from 'valibot';
 
 import { ajv } from '$lib/adapters/ajv.js';
 
@@ -22,7 +22,6 @@ import { type } from 'arktype';
 
 import { typebox } from '$lib/adapters/typebox.js';
 import { Type } from '@sinclair/typebox';
-import { schemaShape } from '$lib/jsonSchema/schemaShape.js';
 //import { TypeCompiler } from '@sinclair/typebox/compiler';
 
 ///// Test data /////////////////////////////////////////////////////
@@ -50,13 +49,14 @@ const validData = {
 
 /**
  * Input data to superValidate
- * Should give error on email, tags, tags[1] and score
+ * Should give error on email, tags and tags[1]
  * Score is left out, to see if defaults are added properly.
  */
 const invalidData = { name: 'Ok', email: '', tags: ['AB', 'B'] };
 
 /**
  * What should be returned when no data is sent to superValidate
+ * Should give error on email and tags
  */
 const defaults = { name: '', email: '', tags: [], score: 0 };
 
@@ -111,13 +111,14 @@ describe('Arktype', () => {
 	const schema = type({
 		name: 'string',
 		email: 'email',
-		tags: '(string>=2)[]>=2'
+		tags: '(string>=2)[]>=3',
+		score: 'integer>=0'
 	});
 
 	const errors = {
 		email: "email must be a valid email (was '')",
-		//tags: 'tags must be at least 2 characters (was 0)',
-		tags1: 'tags/1 must be at least 2 characters (was 1)'
+		tags: /tags must be at least 3 characters \(was [02]\)/
+		//tags1: 'tags/1 must be at least 2 characters (was 1)'
 	};
 
 	schemaTest(() => arktype(schema, { defaults }), errors, false);
@@ -125,18 +126,18 @@ describe('Arktype', () => {
 
 /////////////////////////////////////////////////////////////////////
 
-/*
 describe('Valibot', () => {
 	const schema = object({
 		name: string(),
 		email: string([email()]),
-		tags: array(string([minLength(2)]), [minLength(2)])
+		tags: array(string([minLength(2)]), [minLength(3)]),
+		score: number([integer(), minValue(0)])
 	});
 
 	const errors = {
 		email: 'Invalid email',
 		tags: 'Invalid length',
-		tags0: 'Invalid length'
+		tags1: 'Invalid length'
 	};
 
 	schemaTest(() => valibot(schema, { defaults }), errors, false);
@@ -152,20 +153,20 @@ describe('ajv', () => {
 			email: { type: 'string', format: 'email' },
 			tags: {
 				type: 'array',
-				minItems: 2,
-				items: { type: 'string', minLength: 2 },
-				default: ['A'] as string[]
-			}
+				minItems: 3,
+				items: { type: 'string', minLength: 2 }
+			},
+			score: { type: 'integer', minimum: 0 }
 		},
-		required: ['name', 'email', 'tags'] as string[],
+		required: ['name', 'email', 'tags', 'score'] as string[],
 		additionalProperties: false,
 		$schema: 'http://json-schema.org/draft-07/schema#'
 	} as const;
 
 	const errors = {
 		email: 'must match format "email"',
-		tags: 'must NOT have fewer than 2 items',
-		tags0: 'must NOT have fewer than 2 characters'
+		tags: 'must NOT have fewer than 3 items',
+		tags1: 'must NOT have fewer than 2 characters'
 	};
 
 	schemaTest(() => ajv(schema), errors, true);
@@ -215,27 +216,25 @@ describe('Zod', () => {
 
 	schemaTest(() => zod(schema), errors, true);
 });
-*/
 
 ///// Test function for all validation libraries ////////////////////
 
 function schemaTest(
 	adapter: () => ValidationAdapter<Record<string, unknown>>,
-	errors: { email: string; tags?: string; tags1?: string },
+	errors: { email: string | RegExp; tags?: string | RegExp; tags1?: string | RegExp },
 	testConstraints: boolean
 ) {
-	function expectedErrors(addTag1 = true) {
+	function expectErrors(errorMessages: Record<string, any>, hasTagContentErrors = true) {
 		// eslint-disable-next-line @typescript-eslint/no-explicit-any
-		const output: Record<string, any> = {
-			email: [errors.email]
-		};
+		expect(errorMessages.email).toMatch(errors.email);
 
-		if (addTag1 && errors.tags1) output.tags = { '1': [errors.tags1] };
 		if (errors.tags) {
-			if (!output.tags) output.tags = {};
-			output.tags._errors = [errors.tags];
+			expect(errorMessages?.tags?._errors?.[0] ?? '').toMatch(errors.tags);
 		}
-		return output;
+
+		if (hasTagContentErrors && errors.tags1) {
+			expect(errorMessages?.tags?.['1']?.[0] ?? '').toMatch(errors.tags1);
+		}
 	}
 
 	it('with schema only', async () => {
@@ -250,9 +249,9 @@ function schemaTest(
 
 	it('with schema only and initial errors', async () => {
 		const output = await superValidate(adapter(), { errors: true });
-		const errors = expectedErrors(false);
-
-		expect(output.errors).toEqual(errors);
+		// Expect default value errors, which means that tags[1] should not exist,
+		// the error is only for the array length.
+		expectErrors(output.errors, false);
 		expect(output.valid).toEqual(false);
 		expect(output.data).not.toBe(defaults);
 		expect(output.data).toEqual(defaults);
@@ -260,9 +259,9 @@ function schemaTest(
 		if (testConstraints) expect(output.constraints).toEqual(expectedConstraints);
 	});
 
-	it('with incorrect test data', async () => {
+	it('with invalid test data', async () => {
 		const output = await superValidate(invalidData, adapter());
-		expect(output.errors).toEqual(expectedErrors());
+		expectErrors(output.errors);
 		expect(output.valid).toEqual(false);
 		expect(output.data).not.toBe(invalidData);
 		// Defaults and incorrectData are now merged